src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiofiles==23.2.1
    # via quart
aiohttp==3.9.3
    # via quartapp (pyproject.toml)
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.2.0
    # via
    #   httpx
    #   openai
    #   watchfiles
attrs==23.2.0
    # via aiohttp
azure-core==1.30.0
    # via azure-identity
azure-identity==1.15.0
    # via quartapp (pyproject.toml)
blinker==1.7.0
    # via
    #   flask
    #   quart
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   quart
    #   uvicorn
cryptography==42.0.2
    # via
    #   azure-identity
    #   msal
    #   pyjwt
distro==1.9.0
    # via openai
flask==3.0.2
    # via quart
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gunicorn==21.2.0
    # via quartapp (pyproject.toml)
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   uvicorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
httpcore==1.0.2
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.26.0
    # via openai
hypercorn==0.16.0
    # via quart
hyperframe==6.0.1
    # via h2
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.1.2
    # via
    #   flask
    #   quart
jinja2==3.1.3
    # via
    #   flask
    #   quart
markupsafe==2.1.5
    # via
    #   jinja2
    #   quart
    #   werkzeug
msal==1.26.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.1.0
    # via azure-identity
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
openai==1.12.0
    # via quartapp (pyproject.toml)
packaging==23.2
    # via
    #   gunicorn
    #   msal-extensions
portalocker==2.8.2
    # via msal-extensions
priority==2.0.0
    # via hypercorn
pycparser==2.21
    # via cffi
pydantic==2.6.1
    # via
    #   openai
    #   quart-schema
pydantic-core==2.16.2
    # via pydantic
pyhumps==3.8.0
    # via quart-schema
pyjwt[crypto]==2.8.0
    # via msal
python-dotenv==1.0.1
    # via
    #   quartapp (pyproject.toml)
    #   uvicorn
pyyaml==6.0.1
    # via
    #   quartapp (pyproject.toml)
    #   uvicorn
quart==0.19.4
    # via
    #   quart-schema
    #   quartapp (pyproject.toml)
quart-schema[pydantic]==0.19.1
    # via quartapp (pyproject.toml)
requests==2.31.0
    # via
    #   azure-core
    #   msal
six==1.16.0
    # via azure-core
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.2
    # via openai
typing-extensions==4.9.0
    # via
    #   azure-core
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.0
    # via requests
uvicorn[standard]==0.27.1
    # via quartapp (pyproject.toml)
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
werkzeug==3.0.1
    # via
    #   flask
    #   quart
    #   quartapp (pyproject.toml)
wsproto==1.2.0
    # via hypercorn
yarl==1.9.4
    # via aiohttp
